[Refactor] myrecipe package
