+++ app/src/main/java/com/doctor/yumyum/presentation/ui/myrecipe/viewmodel/MyPageFilterViewModel.kt
@@ -1,0 +1,7 @@
+package com.doctor.yumyum.presentation.ui.myrecipe.viewmodel
+
+import com.doctor.yumyum.common.base.BaseViewModel
+
+class MyPageFilterViewModel : BaseViewModel() {
+
+}
